requirements.txt
faster-whisper
streamlit
streamlit-authenticator
pyyaml
openai
azure-cognitiveservices-speech
pytesseract
gTTS
httpx
pytube
deep-translator
srt
psutil
elevenlabslib
pytz